[storage-aws] CopierTest extract from S3StorageServiceSuite
