+++ storage-aws/src/test/scala/net/kemitix/thorp/storage/aws/S3StorageServiceSuite.scala
@@ -4,21 +4,10 @@ import java.time.Instant
 import java.time.temporal.ChronoUnit
 import java.util.Date
 
-import com.amazonaws.services.s3.model.{
-  AmazonS3Exception,
-  CopyObjectResult,
-  ListObjectsV2Result,
-  S3ObjectSummary
-}
+import com.amazonaws.services.s3.model.{ListObjectsV2Result, S3ObjectSummary}
 import net.kemitix.thorp.console.MyConsole
 import net.kemitix.thorp.core.Resource
-import net.kemitix.thorp.domain.StorageQueueEvent.{
-  Action,
-  DoNothingQueueEvent,
-  ErrorQueueEvent
-}
 import net.kemitix.thorp.domain._
-import net.kemitix.thorp.storage.aws.S3ClientException.S3Exception
 import org.scalamock.scalatest.MockFactory
 import org.scalatest.FreeSpec
 import zio.Runtime
@@ -79,77 +68,6 @@ class S3StorageServiceSuite extends FreeSpec with MockFactory {
       runtime.unsafeRunSync {
         storageService
           .listObjects(Bucket("bucket"), RemoteKey("prefix"))
-      }.toEither
-  }
-
-  "copier" - {
-    val bucket    = Bucket("aBucket")
-    val sourceKey = RemoteKey("sourceKey")
-    val hash      = MD5Hash("aHash")
-    val targetKey = RemoteKey("targetKey")
-    "when source exists" - {
-      "when source hash matches" - {
-        "copies from source to target" in {
-          val event    = StorageQueueEvent.CopyQueueEvent(sourceKey, targetKey)
-          val expected = Right(event)
-          new AmazonS3ClientTestFixture {
-            (fixture.amazonS3Client.copyObject _)
-              .when()
-              .returns(_ => new CopyObjectResult)
-            private val result =
-              invoke(bucket, sourceKey, hash, targetKey, fixture.storageService)
-            assertResult(expected)(result)
-          }
-        }
-      }
-      "when source hash does not match" - {
-        "skip the file with an error" in {
-          val event    = DoNothingQueueEvent(targetKey)
-          val expected = Right(event)
-          new AmazonS3ClientTestFixture {
-            (fixture.amazonS3Client.copyObject _)
-              .when()
-              .returns(_ => null)
-            private val result =
-              invoke(bucket, sourceKey, hash, targetKey, fixture.storageService)
-            assertResult(expected)(result)
-          }
-        }
-      }
-      "when client throws an exception" - {
-        "skip the file with an error" in {
-          new AmazonS3ClientTestFixture {
-            private val expectedMessage = "The specified key does not exist"
-            (fixture.amazonS3Client.copyObject _)
-              .when()
-              .throws(new AmazonS3Exception(expectedMessage))
-            private val result =
-              invoke(bucket, sourceKey, hash, targetKey, fixture.storageService)
-            result match {
-              case Right(
-                  ErrorQueueEvent(Action.Copy("sourceKey => targetKey"),
-                                  RemoteKey("targetKey"),
-                                  e)) =>
-                e match {
-                  case S3Exception(message) =>
-                    assert(message.startsWith(expectedMessage))
-                  case _ => fail("Not an S3Exception")
-                }
-              case e => fail("Not an ErrorQueueEvent: " + e.toString)
-            }
-          }
-        }
-      }
-    }
-    def invoke(
-        bucket: Bucket,
-        sourceKey: RemoteKey,
-        hash: MD5Hash,
-        targetKey: RemoteKey,
-        storageService: S3StorageService
-    ) =
-      runtime.unsafeRunSync {
-        storageService.copy(bucket, sourceKey, hash, targetKey)
       }.toEither
   }
